completed authorization - (protecting routes and verifying current user)

+++ routers/posts.py
@@ -9,6 +9,8 @@ import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
 
+from auth import CurrentUser
+
 router = APIRouter()
 
 # api process
@@ -32,7 +34,7 @@ async def get_post_api(post_id: int, db: Annotated[AsyncSession, Depends(get_db)
 
 # update a post (put)
 @router.put("/{post_id}", response_model=PostResponse)
-async def updateFull_post_api(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def updateFull_post_api(post_id: int, current_user: CurrentUser, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
 
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id == post_id))
     post = result.scalars().first()
@@ -40,15 +42,11 @@ async def updateFull_post_api(post_id: int, post_data: PostCreate, db: Annotated
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 
-    if post_data.user_id != post.user_id:
-        result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    if post.user_id != current_user.id :
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -57,7 +55,7 @@ async def updateFull_post_api(post_id: int, post_data: PostCreate, db: Annotated
 
 # update a post (patch)
 @router.patch("/{post_id}", response_model=PostResponse)
-async def updatePartial_post_api(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def updatePartial_post_api(post_id: int, current_user: CurrentUser, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
 
     result = await db.execute(select(models.Post).options(selectinload(models.Post.author)).where(models.Post.id == post_id))
     post = result.scalars().first()
@@ -66,6 +64,9 @@ async def updatePartial_post_api(post_id: int, post_data: PostUpdate, db: Annota
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 
     update_post_dict = post_data.model_dump(exclude_unset=True)
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
 
     for field, value in update_post_dict.items():
         setattr(post, field, value)
@@ -77,14 +78,9 @@ async def updatePartial_post_api(post_id: int, post_data: PostUpdate, db: Annota
 
 # create a post
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post_api(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def create_post_api(post: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
 
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
-
-    new_post = models.Post(title=post.title, content=post.content, user_id=post.user_id)
+    new_post = models.Post(title=post.title, content=post.content, user_id=current_user.id)
     db.add(new_post)
     await db.commit()
     await db.refresh(new_post, attribute_names=["author"])
@@ -93,13 +89,16 @@ async def create_post_api(post: PostCreate, db: Annotated[AsyncSession, Depends(
 
 # delete a post
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post_api(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post_api(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
 
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found.")
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this post")
 
     await db.delete(post)
     await db.commit()
